--- .gitmodules
@@ -1,6 +1,0 @@
-[submodule "lab5"]
-	path = lab5
-	url = https://github.com/dimakyba/Lab5_2nd_sem
-[submodule "lab3"]
-	path = lab3
-	url = https://github.com/dimakyba/Lab3_PaAL
