+++ README.md
@@ -1,11 +1,3 @@
 TO-DO
 
-- [ ] barni
-
-- [ ] Äâà âåêòîðè ââàæàþòü ð³âíèìè, ÿêùî âîíè êîë³íåàðí³ ³ ìàþòü ( пофіксити кирилицю)
-
-- [ ] generics
-
-- [ ] vlad
-
-- [ ] exception handler
+- [ ] кирилиця
