--- lab5/.gitignore
@@ -1,3 +1,0 @@
-bin
-obj
-.vscode
