--- lab5/Orders.json
@@ -1,7 +1,0 @@
-﻿[
-  {
-    "SenderAccount": "\u0000\u0000\u0000\u0000",
-    "ReceiverAccount": "\u0000\u0000\u0000\u0000",
-    "SumInCents": 1488
-  }
-]
