--- lab5/README.md
@@ -1,3 +1,0 @@
-TO-DO
-
-- [ ] кирилиця
